Вынес константы в mock.js
