+++ js/create-offer.js
@@ -1,20 +1,10 @@
 import {getRandomNumber} from './get-random-number.js';
+import {OFFER_TYPE, OFFER_TIMES, OFFER_FEATURES, OFFER_PHOTOS, OFFER_TITLE, OFFER_DES} from './mock.js';
 
-const OFFER_TYPE = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
-const OFFER_TIMES = ['12:00', '13:00', '14:00'];
-const OFFER_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
-const OFFER_PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-const OFFER_TITLE = ['молодых котов', 'котов в расцвете лет', 'котих', 'кошек', 'котов с людьми'];
-const OFFER_DES = ['своим великолепием', 'своей ценой и качеством', 'своими размерами', 'воображение'];
 const OFFER_LAT_FROM = 35.65000;
 const OFFER_LAT_TO = 35.70000;
 const OFFER_LNG_FROM = 139.70000;
 const OFFER_LNG_TO = 139.80000;
-
 
 const createOffer = () => {
   const LAT = getRandomNumber(OFFER_LAT_FROM,OFFER_LAT_TO);
